Fold old cli.py into cli_dispatcher.py, move cli_dispatcher to cli.py
cli.py had very little code, cli_dispatcher doesn't in globals do any bad state mutation. Even if it had, we were importing it fully which would have picked that up anyways.
